refactor: MultiProofTargetsV2::from_state
